Adds support for "knots", a pseudo-DAG.
This allows a child to have two parents, but no more.

The approach is to try and reduce to the tree case when
no knots are present, or within O(1) of it (incrementally,
so without "are there any knots?" type conditionals).

This is mostly done by adding fallback knot-checks where
a single-parent/single-child lookup fails.

Secondary priority is limiting the data footprint, and we work
with an assumption that the number of knots are small in
comparison to the graph size. Hence, we have a knot component,
and will search across knots rather than iterating through
entities/children/parents where possible. The knot component
is essentially a pair of children ("left" and "right") who
retain their respective parents, and appear in the respective
child linked-lists. The knot component is only tracked from
the archetype column.

Limitations: detaching currently has only two modes available,
IGNORE, which does not consider the possibility of nodes below
the detached node tying the substructure back to the original
tree through a knot, and PRUNE, which removes any such shared
descendents from the tree at the same time. This increases the
operations required for a detach, as a descendent search is
(in the standard PRUNE case) required.